Refactor the way actions are added
* Actions are added via TransientActionFactory and not directly in CucumberReportPublisher
* This will enable the ProjectAction to also show up in workflow/pipeline jobs
* Move actions to separate package and deprecate old implementations
* Simplify ProjectAction by using lastCompletedBuild when computing path
* Rename SafeArchiveServingAction to CucumberReportRunAction
