+++ src/main/java/net/masterthought/jenkins/CucumberReportBaseAction.java
@@ -1,9 +1,12 @@
 package net.masterthought.jenkins;
 
 import hudson.model.Action;
-
 import net.masterthought.cucumber.ReportBuilder;
 
+/**
+ * @deprecated functionality moved to action package.
+ */
+@Deprecated
 public abstract class CucumberReportBaseAction implements Action {
 
     protected static final String ICON_NAME = "/plugin/cucumber-reports/icon.png";
